returned the static callback classes and made them private
